1st work

+++ fdf.h
@@ -66,9 +66,9 @@ typedef struct	s_env
 	int			scale_x;
 	int			scale_y;
 	int			scale_z;
-	int		angle_x;
-	int		angle_y;
-	int		angle_z;
+	double		angle_x;
+	double		angle_y;
+	double		angle_z;
 	double			depth;
 	int			scale;
 }				t_env;
